Queue watch fix, More comments
When I added the debugging message statements in the queue watch
there was a bug that wasn't noticed. This caused it not to display
some hits in the queue. Had to make sure the debugging statements
was surrounded in a bracket for the if statement.

Some more jsdoc comments added to class files.
Changed some function calls because arrow functions are binded to this
Version number changed to 0.8.2