Grouping fixes and more commenting.
I fixed it so groupings will toggle off if all their panda's are off.
I fixed it so groupings will toggle on if all their panda's are on.
Added a function to convert database ID to my ID in panda class.
Took out some calls to functions in modal class.
Added more comments for classes.

Forgot to turn off database deletion at startup when I was testing
in the panda crazy openDB method.