Fixed errors and changed callbacks.
There was a problem at the beginning if a person was logged off
it can show the logged off modal before loading modal closes.
Made sure the loading modal closes first before starting queue.

When I changed the way callbacks was used it broke the popup
cross domain detection so it wouldn't close the logged off modal
after the popup was closed or when a person logs back in.
Changed it so it checks the popup closed value and will tell queue
that it should be logged back in.